[v0.4.2] Fixed data loading delay, added file caching
Previously there was a bug where when you joined your data wouldnt get loaded until you spoke once, this is fixed.
Additionally, there is now a config option for your system information and member list to be cached to disk so that you don't have to wait on the API to return your information.